docker/python_backup_db/requirements.txt
psycopg2-binary>=2.9
pg_dump